requirements.txt
blinker
click
colorama
Flask
itsdangerous
Jinja2
joblib
MarkupSafe
numpy
pandas
python-dateutil
pytz
scikit-learn
scipy
six
threadpoolctl
tzdata
Werkzeug
streamlit
python-docx
PyPDF2
